Transformer/requirements.txt
torch
numpy
pandas
omegaconf
tqdm
scikit-learn
PyYAML
tensorboard
matplotlib
pillow